Added to and reformatted test report; Fix permissions; remove 24hr adjust; Added sound note; remove linux tests.
 Changes to be committed:
	modified:   TESTREPORT

1. Added test result.

2. Reformatted.

	modified:   linux/services.c

1. Changed permission checks to negative.

2. Removed 24hr adjustment, I don't recall what this case was for. It was
improperly encoded in any case, and it can be put back.

	modified:   linux/sound.not

1. Added note. See notes.

	modified:   tests/services_test.c

1. Ifdef'ed attribute tests, which Linux cannot perform.

Notes:

I really can't find the ALSA inputs and outputs needed to perform tests because
the names of devices is so obscure. I think we need to switch to device
descriptions. I guess that ALSA actually needs these device descriptions, so
they would be maintained internally.