Made config.c line ending tolerant.
 Changes to be committed:
	modified:   utils/config.c

1. Created new routine, fgetsale(), fgets() "all line endings".

2. Change fgets() call to above.

Notes:

Config is the only thing in the petit-ami fileset that reads text,
so this effectively makes petit-ami portable for OS line endings.

It could happen that the fgets() be taken back into stdio.c. This would need
be accompanied by a full audit of compatability and if stdio.c can actually
be made line ending toleratant. In any case, it would not solve the current
issue, because we use glibc, not our stdio.c.