Turn off auto background fills; handle background fills for buffered mode.
 Changes to be committed:
	modified:   linux/graphics.c

1. In opnwin(), change to XCreateWindow() call so that we can turn off automatic
background fills, it causes flashing and is the incorrect color in any case (we
did the same thing in Windows).

2. Added rectangle type.

3. Added setrect() set rectangle values, intersect() check intersection of
rectangles, zerorect() check null rectangle, and subrect() find subtraction
elements of two rectangles. This stuff came from edesign.

4. In xwinevt() buffer automatic handling, process the expose event in terms of
intersection of the buffer area with the expose rectangle, then find the
outlying parts of the expose rectangle after subtracting the buffer rectangle.
This is optimized handling of the expose event.

Notes:

1. Restore is going to need the same treatment.

2. The foreground color needs restoring.