Fill pa_getsiz[g].
 Changes to be committed:
	modified:   linux/graphics.c

1. Removed EVTPOL define, it does not work.

2. Added "standard character sizes" from Windows. This is needed for getsiz()
(character version of getsiz()).

3. Added dispxg and dispyg variables to win structure. See notes.

4. In opnwin(), set the dispxg and dispyg variables.

5. In xwinevt(), on the ConfigureNotify event, update dispxg and dispyg.

6. Removed the EVTPOL alternative. There is no way for the graphics module to
function using that method, so remove.

7. Set dispxg and dispyg in sizbufg(). See notes.

8. Fill out getsizg() and getsiz() functions.

Notes:

getsiz[g] normally would determine the display size using XGetWinowAttributes.
The issue is apparently that XWindows does not actually set the window
characteristics until a round trip communication with the window manager, so it
is possible that a call to getsiz[g]() would return incorrect information if
executed soon after a set. The workaround was to track the display size
internally from events, but that, too, is too late, so if we resize the display,
we set it immediately.