Refactor getsiz/sizbuf; gather window parameters.
 Changes to be committed:
	modified:   linux/graphics.c

1. Remove dispxg and dispyg parameters from win struct.

2. Added pfw (frame extra width), pfh (frame extra height), cwox (client offset
x) and cwoy (client offset y). These are parameters we need for later windows
calls.

3. Change border parameters to XCreateWindow() to 0. Turns out they are ignored
and set to zero in any case.

4. Remove dispxg and dispyg from opnwin().

5. opnwin() gets pfw, pfh, cwox and cwoy parameters by finding the parent
window, which is assumed to be the frame window, and caluculating the difference
between the client and the frame parent.

6. Remove dispxg and dispyg from xwinevt().

7. Refactor sizebufg() to wait for ConfigureNotify event.

8. Refactor getsizg() to get the demensions of its parent, which is assumed to
be the window frame.

Notes:

1. The solution to the XConfigureWindow() to XGetWindowAttributes() delay is to
hold for the ConfigureNotify message that indicates that XConfigureWindow() has
taken effect. This is loaded with problems. There could be an ConfigureNotify
already in the queue, or we could discard an important message like a keystroke
while looking for the ConfigureNotify message. I am looking for answers.

2. We assume the parent of the window is it's frame. We could do better at
verifying that this is true.