Refactor frame(); Fill out sizable() and sysbar().
 Changes to be committed:
	modified:   linux/graphics.c

1. Put in newer, better definitions of Motif mode sets.

2. Refactored frame() to use new motif definitions.

3. Filled in sizable() function. See notes.

4. Filled in sysbar() function. See notes.

Notes:

The original _NET_WM_WINDOW_TYPE properties from the "extended window manager
hints specification" does not really work (after trying out all of the possible
types). As a result, most people who have to do real work still use the Motif
hints. Many, or even most, window managers implement the "old" motif properties,
however, that implementation is spotty, at least on the GNOME/GDM3 window
manager I tested this on.

Thus frame() and sysbar() work, but sizable does not. This is not a big loss,
since the most common operation is to turn off the entire frame with frame().
The PA manual indicates that all of these function are optional (and that was
written about 20 years ago!). However, the reality is that lack of frame() would
break a lot of stuff (no widgets).

In any case, I made the best effort to set up sizable() to work if a later
window manager were to implement the Motif modes properly. Of course, its always
possible a future WM could break any or all of them.