Refactor for input X events (not checked!); added input queuing for XEvents.
 Changes to be committed:
	modified:   linux/graphics.c

1. Declaration for event queue structures.

2. Add event queue list.

3. Add event queue free list.

4. Add getxevt() to get a new/freed event queue entry.

5. Add putxevt() to release a used event queue entry.

6. Add quexevt() to enter a new event into the queue.

7. Add dequexevt() to dequeue an event from the queue.

8. Refactor waitxevt() to enqueue the events instead of dump them.

9. Form new routine from xwinget() called xwinprep() that does just the
processing on the xevent, not checking the XWindow input. This does diagnostic
dumps on the event, determination of which window the event is for, and dispatch
the event via xwinevt().

10. Refactor xwinget() to call xwinprep().

11. In ievent(), check if the input event queue is empty, and dequeue from it if
not, then process that message.

Notes:

This will solve the problem where we need to accept signaling from the Windows
manager for things like configuration and mapping, but still need to preserve
events like exposure. This issue, noted previously, was found to occur when
opening more than one window at once. The solution is we look ahead in the
XEvent input, and then place the events in a queue for pa_event() to pick up.
This mechanisim allows us to look at all events up to the present, but preserve
them to be read by pa_event() at a later time. Think of it like an S curve (S on
it's side), where the first curve of the events comes to the waitxevt() routine
that examines flags from the window manager, then puts them back into a queue,
then the next part of the S brings them back into the pa_event() handler.