Removed diagnostic; expanded pa event diagnostic; Added button press capability.
 Changes to be committed:
	modified:   linux/graphics.c

1. Removed diagnostic from peekxevt().

2. Added window id to PA event diagnostic.

3. Added button pressed ability to buttong() demo.

4. Turn off auto in buttong().

5. Set sign font in buttong().

6. Set bold in buttong().

Notes:

Button demo completed satisfactorily. Next step is to break this code off into a
widgets.c module, probably in a new directory "portable" or somesuch, since it
is not OS specific code.

The button study code is also the basis for the menu capability, since it will
use very similar methods. Thus, buttong() is the prototype widget implementation
in PA.