Add sendevent(); Add pa event queuing; Added button event to buttong().
 Changes to be committed:
	modified:   include/graphics.h

1. Add pa_sendevent() definition.

	modified:   linux/graphics.c

1. Add new error for no input side of window.

2. Add PA event queue structure define.

3. Add paqfre, the PA free queue entries.

4. Add paqevt, the PA queue entries.

5. Changed quexevt() name to enquexevt().

6. Added getpaevt(), get a new PA queuing entry.

7. Added putpaevt(), free a PA queuing entry.

8. Added enquepaevt(), place PA event in queue.

9. Added dequepaevt(), remove PA event from queue.

10. Added sendevent(), routine that determines window id from passed file a
constructs an event from the given and queues it. See notes.

11. Added buttong() ability to send a button event.

12. Clear PA free and queue lists.

Notes:

Created a queue to allow PA events from any caller to be put at the front of the
pa_event() input. This means that events can be placed into the input at any
time, and they will wait until pa_event() is called. Placing new events in the
event override chain is possible, but not as good because you don't know what
handlers in the chain would be skipped. Also, this works for sending events
into the input events at any time.

Note the dequeue in pa_event() is not implemented. Next fix.