Add buried wid facillity for widgets.
 Changes to be committed:
	modified:   include/graphics.h

1. Added define for getwid() call, get buried wid.

	modified:   linux/graphics.c

1. Added "no more wids" error for buried wids. The user dosen't need to know or
care what kind of wid ran out.

2. Double and add one for the xltwin[] table. This allows for negative wids,
plus one for the zero table entry. This allows us to do a direct index to the
table by xltwin[wid+MAXFIL], with 0 at the middle.

3. Do several xltwin[] lookup converts to xltwin[wid+MAXFIL] form.

4. Add getwid() routine. This allocates wids from -1 to -MAXFIL+1, which I just
realized is unnecessary, since wid >=0 will do it.

5. Create separate init for xltwin[] table, since it no longer is MAXFIL size.

Notes:

The purpose of the buried wid system is to create wids that are automatically
allocated and can never overlap the normal (positive) wids. If, for example, we
created special wids at 1000 or more, we could never garantee that the window id
numbers required for the client could not reach that number. wids are on a
client "honor" system. The client maintains the wid numbers as it pleases, it
could event allocate them automatically itself. The real limit on wids is
arbitrary, and could even be N if we feel like using lists someday.

Instead, using negative numbers, we insure the automatically allocated wids and
the client wids will never overlap.

Auto allocated wids are needed by the widget package, since it needs wids that
the client never would use, and also the widget package does not need to track.

I created the getwid() call as an open, general call in case the client wants
to use an automatic allocator without making one itself. This has the advantage
that it can mix automatic allocates and static allocates. Also the client
can make its own widgets.