Fixed definition of menuspcy; remove double line menu ability; Fix mettrk() bug.
 Changes to be committed:
	modified:   linux/graphics.c

1. In opnwin(), fix bad definition of menuspcy.

2. In actmenu, remove ability to divide meny into multiple lines. See notes.

3. Fix bug in mettrk() where branch was not getting initialized.

Notes:

Multiple line menus, which I call "folding" appear to be not used in Gnome.
Several examples were examined, including Glade examples, and the menu is either
clipped at right or the window reduction is limited to the x size of the menu.
Either indicates folding is not implemented, and the code as is will support
either.