Added master/subclient window architecture.
 Changes to be committed:
	modified:   linux/graphics.c

1. Added xmwhan, master XWindow handle to windows structure.

2. Changed fndevt() to also check the master window handle.

3. winvis() maps the master window, then the subclient window.

4. New routine, createwindow(), creates a backgroundless window with the given
dimensions and selects the events from it.

5. opnwin() now creates both the master window and the subclient window.
The borders measurement is done on the master window.

6. setsizg() now references the master window, which is the outtermost
framed window.

7. setposg() now references the master window.

Notes:

The purpose of the master/subclient structure is that the client can be
relocated within the master window. Previously, all windows were masters and
the client was fixed within the window. Now, the subclient can be resized and
moved within the master window. All draws go to the sublclient, and the master
is now just a container for other windows.

This allows the subclient to be moved down and a menu bar window created at the
top of that.