Added subclient configure tracking; add display flush; lock out master exposes.
 Changes to be committed:
	modified:   linux/graphics.c

1. Add display size tracking variables dw and dh to window structure.

2. In opnwin(), set up tracking variables to initial maxxg and maxyg.

3. In xwinevt(), lock out master expose events. This didn't address problem in
particular, its just that the master is a container and should never refresh
anything.

4. Added XFlush() to the expose handler in xwinevt(). See notes.

5. In xwinevt(), added display size tracking. This is a more definiate method of
tracking than previous (we didn't track display sizes in buffered mode). We have
to reject any action on configures with the same parameters, it will cause the
sync code with the window manager to lock up, since the WM does not give answers
for null configures.

6. In xwinevt(), send configures to the master on to the client. IE., resizes
to the master resize its children as well. Note that we have not accounted for
the menu bar here yet.

Notes:

It is very disturbing that we needed to flush the XWindows buffer using
XFlush(), but I saw issues with redraw if I did repeated fast resizes on the
window, specifically black unrefreshed areas. With some added diagnostics, I saw
the correct events flowing, with correct sizing data, but it simply looks like
the draws aren't happening. Note as well that mousing over the incorrect window
will fix it. It is literally like draws are being held indefinitely until other
action is taken on the window.