Supress repeated window size operations.
 Changes to be committed:
	modified:   linux/graphics.c

1. In pa_setsizg(), check if the size operation is repeated using the dw and wh
parameters of the window structure created previously. See notes.

Notes:

This is the same improvement that went into setposg(), which is that if the
window manager sees a repeated operation, it just drops it, and does not make
the change via ConfigurationNotify. This means we hang up looking for it.
So we supress the duplicate operation ourselves.