Added floating menu capability.
 Changes to be committed:
	modified:   linux/graphics.c

1. Added evtfil field to widget data, shows which output window file to post to
regardless of parent.

2. Added out2inp(), returns the input file for any output file attached to a
window.

3. Added fndloc(), translates client coordinates to desktop coordinates (does
not work as yet).

4. Move openwmenu() up.

5. Refactor openmenu() to take input file.

6. Add fltmen(), present floating menu.

7. In menu_event(), activate a floating menu on menu branches.

8. Refactor actmenu() to establish input side file for menu windows.

9. In mettrk(), establish event post window.

10. createmenu() passes down event post window.

11. pa_menu() passes event post window.

12. Fix pa_frame() to operate on master window.

13. Fix pa_sizable() to operate on master window.

14. Fix pa_sysbar() to operate on master window.

Notes:

1. No teardown of floating window stack as yet.

2. The presentation of floating window stacks is UNBELIEVABLY SLOW. This has to
get fixed for this system to be usable. I believe it is primarily because of the
lack of delayed window display, since all the configurations of menu buttons has
to be done onscreen. Ultimately this could cause us to bring the entire menu
system down to X level, which would be unfortunate, not only because it would
make the code more obscure, but also because it calls into question the idea of
creating widgets at the PA level.

3. The position of floating menus needs to be worked out.

4. The floating menu buttons are all shadowed. This is window manager nonsense.
Its obvious both that GTK and others can turn this off, just as it is obvious
that GTK can present menu buttons as screen level objects.