Converted structures to modern form; added window structure recycling; add menus to recycle tracking; add focus events.
 Changes to be committed:
	modified:   linux/graphics.c

1. Several structure declarations were converted to pointer first form, which is
cleaner.

2. Added winfre, windows structure free list.

3. Added menu entries tracking.

4. Added getwin(), get windows structure entry.

5. Added putwin(), put windows structure entry to free.

6. Added focus events to input mask in createwindow(). This is to study if we
want to import focus events (jury still out on this one).

7. Added menu entry tracking in getmet().

8. Added menu tracking inits.

9. Added win free init.

Notes:

memory tracking was put in to solve an issue that turned out not to have
anything to do with memory issues, but it is a good facillity to have, so I am
maintaining it. It could probably use exporting to windows graphics.c.