Implement backing frames for pulldown menus.
 Changes to be committed:
	modified:   linux/graphics.c

1. Add frame field to metrec, this is a metptr that holds the pulldown frame
entry.

2. Add frm field to metrec, this boolean types the metrec as holding a frame.

3. In (the highly misnamed) fltmen(), present a pulldown frame and make the menu
windows children of that.

4. Remove frames in remmen().

5. In menu_event(), remove the checks for bar, that was an initialization error.

6. In menu_event(), add frame draw.

7. In mettrk(), make m, the client menu parameter, optional by NULL. The fields
it controlled get defaults if not present.

8. In createmenu(), add frame menu tracking entry.

9. In pa_menu(), initialize proper fields for bar entry.

Notes:

We now place a frame window behind every pulldown menu, and the menu buttons in
the pulldown menu are children of that window. This actually means we could go
back and implement full floating menus. I believe that will be an option for
later.