Move linux/XWindows widgets to separate package.
 Changes to be committed:
	modified:   Makefile

1. Added component compile for gnome_widgets.c.

	modified:   linux/graphics.c

1. Remove widget code.

	new file:   portable/gnome_widgets.c

1. Add widget code from linux/graphics.c. Only button[g]() is defined at
present.

Notes:

The Gnome widgets package will be OS indenpendent, even though it is Gnome
specific. It is expected that there will be fairly small differences between
desktops and there will be much commonality. The next widget module is
(apparently) KDE, so we will see. Where it is possible to configure by tables,
I plan to take advantage of that (say, color tables). This remains to be
seen.