Fix focus spelling; add hover events.
 Changes to be committed:
	modified:   include/graphics.h

1. Fix focus events spelling.

2. Add hover events.

	modified:   linux/graphics.c

1. Fix focus events spelling.

2. Generate hover events.

Notes:

Hover events are actually needed to do look and feel in Linux/Gnome. Focus
events are not as useful, and actually can be ignored most of the time, since
the meaning of focus is "you can get key events". You can find out the same
thing by, you know, getting a key event. Hovers are different because although
you can know when the mouse has entered a window, you get a mouse move to that
window, but when the mouse leaves a window, you don't know.

So why include both? Its a good question. The client may want to change states
if the window has focus (widgets do this), but also, there may be an alternative
way to get focus, for instance many systems allow you to tab between fields.

In any case, the portable widget library drove the need for these events, which
I will admit is strange, since it does not use them at present, and works fine
anyways. Its a look and feel thing.