Fix issue with frame timer.
 Changes to be committed:
	modified:   linux/graphics.c

1. In ievent(), set window id for frame timer event.

2. In frametimer(), set window for frame timer request.

3. In init(), get a file entry for the frame timer.

Notes:

The frame timer had to be directed to a particular window. In effect, this
becomes the last window that requested it. This is not an issue in practice,
since different window handlers of the same event queue don't care what window
it is directed to. The frame timer (like other timers) can only be directed to
one window input queue. This may have to be fixed up in the future, if it comes
from a central event like the actual retrace interrupt.