Refactor frametest() to always present a screen.
 Changes to be committed:
	modified:   tests/management_test.c

1. Add new routine, frameinside(), that draws the inside of the frametest.

2. Refactor frametest() to use the new routine, and also use it at the start
of the test. See notes.

Notes:

In Xwindows with GDM3 window manager, size bars cannot be turned off and on
individually. This is alowed by the rules, but frametest() was structured such
that if the reconfigure is a no-op, the display does not change or even print
title, either. This leads the user to believe the test didn't work.

Instead, we always print the frametest contents, regardless of if the operation
worked or not.