Expand the number of joystick axes possible to 6.
 Changes to be committed:
	modified:   include/graphics.h
	modified:   include/terminal.h

1. Introduced new joystick axes variables to the joystick move event.

	modified:   linux/graphics.c

1. Added 3 new joystick tracking variables.

2. Joyevt() tracks the three possible new joystick moves, but discards any over
6.

	modified:   tests/terminal_test.c

1. Modified the joystick test to take up to 6 axes, and also shows only one
joystick at a time, so that any number of joysticks can appear.

Note:

At 6 axes, 6*int = 24 bytes in a message, makes that the largest PA message that
we send. In any case, the joystick I have (Logitech) wastes 2 of the axes with
on/off switches in any case. The other alternative was to switch to a change
only interface like the Linux joystick sends, but that would require the client
track their own joystick states, which is not how the interface was designed.

Thus, the joystick support got limited to what I have on hand, in its own way
a classic programming error.