Added PA event dump ability; refactored events.
 Changes to be committed:
	modified:   linux/terminal.c

1. Added DMPEVT define.

2. Added system event record and types.

3. Added signal mask global.

4. Added dmpevt variable.

5. Added inisevt(), initialize the system events handler.

6. Added addsefid(), add logical file id to system events.

7. Added addsesig(), add signal to system events.

8. Added getsevt(), get next system event or wait for it.

9. Removed schjoy(), getsevt() does that work now.

10. ievent() was refactored extensively to use getsevt() to do the work.

11. PA event dumper was added to pa_event().

12. timer() uses the new method to register the timer fid.

13. In init(), set dmpevt flag.

14. In init(), moved the system event/select() stuff to the above routines.

Notes:

The event mechanisim was refactored so that all of the system events are placed
in a single routine, getsevt(), with several helper routines to initialize the
system and place events of interest within it.

The change was precipitated by the fact that BSD and Mac OS X use a different
mechanisim to include timers into events than the select/pselect call. Thus
separating out the system events into a central routine is preparitory to adding
that different code, either by replacing the getsevt() routine or via #ifs.

Note that the act of receiving a system signal and doing something about it are
completely separate. getsevt() does clear the mask for the event, but nothing
else. Reading the file for the event, clearing the signal, all of that is done
by the caller. system events just tell you something happened, not what to do
about it.

This code is the first stage. Timers are not really abstracted as yet. That will
occur in the next commit.