Add deinit to system events; add timer activate and deactivate.
 Changes to be committed:
	modified:   linux/terminal.c

1. Add deinisevt(), deinitialize system events.

2. Add addsetim(), add system timer event.

3. Add deasetim(), stop system timer event.

4. Modify timer() to use addsetim().

5. Modify killtimer() to use deasetim().

6. Refactor deinit() to use deinisevt().

Notes:

I think the goal is to move the system event handler to another module. Right
now it couples two tables, timer and joystick. I want to refactor so that there
is a system event table that stores the type data, then have getsevt() return
a logical event number. ievent can use that to reverse lookup the logical timer
and joystick interfaces. In that way the table coupling will be removed, and the
system even handler actually gets a bit simpler.