Fix console font size and create settable parameter; Add graphics notes; Add console font size.
 Changes to be committed:
       modified:   linux/graphics.c

1. Change console font size to 11 (see notes).

2. Move console point define to config parameters area.

3. Create conpnt variable.

4. Change console points define to variable.

5. Init console points.

6. Parse console_points parameter from config file.

       modified:   linux/graphics.not

1. Add graphics notes.

       modified:   petit_ami.cfg

1. Add console points parameter.

Notes:

Changed Mac OS X console points to accurate size, 11 points, same as xterm
default. It was 18 points for Linux/Ubuntu because that system gives a crap
answer for the DPM sizes. I don't have an immediate answer for this, but the
problem needs to be solved there instead of on Mac OS X (IE., don't punish
the good for what the bad does).