Fixed various clang nonsense.
Notes:

I like the extra error checking facilities of clang, I don't appreciate
when it starts telling you about coding styles that have nothing to do
with the function of the program. For example, enforcing defaults for
"missing cases" (with the ANSI C standard clearly states are no-ops),
or "add a brace" messages. And finding symbols in duplicate with other
modules when THEY ARE ATTRIBUTED AS STATIC is a BUG in the compiler.
Plain and simple.