Add connectnet program; add gettys test to network_test; Add comment to gettys; Added test compare files.
 Changes to be committed:
	modified:   Makefile

1. Added compile for connectnet program.

	modified:   bin/network_test

1. Added gettys test.

	new file:   network_programs/connectnet.c

1. Added connectnet program, like telnet but can use TLS security.

	modified:   network_programs/gettys.c

1. Added comment.

	new file:   tests/fakemail.txt
	new file:   tests/getpage.txt
	new file:   tests/gettys.txt

1. Added these results compare files, they hold the golden run output data.

Notes:

openssl dumps the certificates into its connection. Not bad, but it also contains
date and time information, and thus does not compare well. Thus we made connectnet.
This program is something of a swiss army knife, like telnet, but can use TLS. So
the obvious question is "why not ssh"? Well turns out that is another, incompatible
protocol. connectnet can talk to any PA program, even a shell.