Set up dummy window; process enable/disable; Reduce redraws; Remove bold; refactor enable; refactor putwidgettext; add sizing routines.
 Changes to be committed:
	modified:   portable/gnome_widgets.c

1. Add "window 0" dummy window.

2. In button_event(), fix check for mouse button 1.

3. In checkbox_draw(), fix check for mouse button 1.

4. In checkbox_draw(), allow enable/disable.

5. Refactor checkbox_event() to reduce redraw.

6. In radiobutton_draw(), allow enable/disable.

7. In radiobutton_event(), refactor to reduce redraw.

8. Remove bold in widget().

9. Fix selectwidget().

10. Fix enablewidget().

11. Add redraw to putwidgettext().

12. Fill in buttonsizg().

13. Fill in checkboxsizg().

14. radiobuttonsizg().

15. In init, establish window 0.

Notes:

The dummy window is used to find font dimensions before actually creating widgets.
The window has buffering and frame turned off, and should not appear on screen
unless the delayed presentation is turned off.

In fact, it does appear briefly in black, but this is because of the problems
with delayed presentation. This will be revisited later.