Fake focus fix.
 Changes to be committed:
	modified:   linux/graphics.c

1. Added child window link and child window chain link to windows structure.
childwin is the top of the child window list for this window, and childlst
allows each window struct to be linked together to form the child window list.

2. In opnwin(), clear the child window list links, and then if the window being
created is a child, push it to the parent's child window list.

3. Added remfocus() routine, expects a window structure pointer and tours the
subwindow tree of that window, clearing any window that has focus.

4. In xwinevt(), refactor to use the remfocus call.

	modified:   linux/graphics.not

1. Remove focus issue. It is fixed now.