Fix event signaling problem.
 Changes to be committed:
	modified:   linux/graphics.c

1. In xwinget(), look for multiple xwindow messages.

Notes:

Identified an issue where multiple xwindows messages are in queue, that only one
select() signal is generated. The solution was to make sure all xwindows
messages are cleared before reentering the select() call
(in system_event_getsevt()). This is a classic signaling/interrupt issue.

Note that the race between xwinget() and system_event_getsevt() is solved
because any messages that could have triggered the select() event that caused
xwinget() to be evoked are retired. Thus any new messages will set a new
select() event, and the cycle repeats.