Second try message stall fix.
 Changes to be committed:
	modified:   linux/graphics.c

1. In xwinget(), we retry the message read if the message processing indicates
that no message was there to process. We do this instead of letting the outer
loop do it, which causes a lock.

Notes:

1. Letting xwinget() leave without a valid message was considered ok because the
outer loop does not exit until it finds a message. However, this can cause the
next message in the queue to not have an event signaled by
system_event_getsevt(), which means it will wait for that signal even though
there is more data.

2. The management_test min/max/norm test does not yet work.