Introduced themetable system.
 Changes to be committed:
	modified:   portable/gnome_widgets.c

1. Added RGB macro to make 32 bit colors from 8 bit components.

2. Added RED(), GREEN() and BLUE() macros to make ratioed colors from 8 bit
components.

3. Define default themetable values.

4. Define themeindex type for the themetable.

5. Add themetable.

6. Add fcolort() and bcolort() routines that take table indexes and set
foreground and background colors.

7. Convert button_draw() to use the themetable.

8. Init the themetable.

Notes:

1. The themetable is a table of 32 bit values. It can be used for anything,
a color, a linewidth, etc.

2. The themetable system uses 8 bit base colors and not 32 bit ratioed colors
as per PA methods. I think it makes sense to do this because:

a. All the current theming systems use that color scheme.

b. There is nothing wrong with a limited selection of preselected colors. The
user still has a full 32*3 colorspace.

The ratioed color system was designed in 2002 and makes lots of sense. There are
already systems capable of more than 8 bit colorspaces.