Subclassed an edit box.
 Changes to be committed:
	modified:   portable/gnome_widgets.c

1. Filled out numselbox_event() (its a no-op at present).

2. In widget(), replaced wp with direct pointer, then we use that in the routine
and assign it back to the caller. This was done because the use of the indirect
pointer was unnecessarily complicating the code.

3. Filled out the numselboxsizg() routine, just with same as editbox for now.

4. Added subclass editbox widget to numselbox widget in numselboxg().

Notes:

This was a test of the ability to subclass a widget. Note that each widget is a
proper window, and has its own relative widget numbering. Thus the subclassed
edit control is simply number 1.