Implement "grab" mode; update widget notes.
 Changes to be committed:
	modified:   portable/gnome_widgets.c

1. Added "grab" flag to widgets.

2. Implemented grab mode in scrollvert_draw() and scrollhoriz_draw(). This sets
the grab flag when starting a mouse drag on the scrollbar, which afterwards
qualifies as having the pointer within the scrollbar. It is cancelled with the
mouse button is released. This means that once the mouse clicks the scrollbar
and starts a drag, the scrollbar will follow the mouse as long as the button is
down. Formerly it was possible for the pointer to outpace the bar, or leave the
widget window, and thus drop the bar. In this way, it only ends when the mouse
releases the button.

	modified:   portable/gnome_widgets.not

1. Register this and the last fix.

	modified:   tests/widget_test.c

Notes:

Still has an issue with losing the scrollbar select highlight.