Removed solved issue from widget notes; added hello world client directory.
 Changes to be committed:
	new file:   Makefile
	new file:   hello.c
	new file:   readme.txt

1. Added client compile example directory. See below.

	modified:   ../portable/gnome_widgets.not

1. Fixed dropbox issue.

Notes:

The hello directory is the first of three directories, hello, helloc, and
hellog. These give sample client compile setups for:

hello  - Plain output model.
helloc - Console output model.
hellog - Graphical output model.

The purpose of this is twofold. First, it gives users of Petit-Ami a starting
setup for there own programs with minimum baggage dragged in from Petit-Ami.
If the user ran a Petit-Ami for their system, all that should be required is to
include the Petit-Ami libraries. Second, it demonstrates what, at minimum, it
takes to compile a program with Petit-Ami, and gives incentive to cut that down
to the barest minimum.

Assuming the barest minimum is just a library include, we have a way to go there.

Issues:

1. We need per-system make commands, so the make needs to select for system. or
does it? Ideally it should not have to.

2. We need to factor out of the make file as much information as possible
(single libraries).