Implemented fix for dropdown/edit box behavior. Involved extending event message system.
 Changes to be committed:
	modified:   include/graphics.h

1. Added reserved event code spaces for each of base system (graphics or
terminal), window manager, widget and user. User goes to infinity.

	modified:   linux/graphics.c

1. Added error message for attempt to vector event from auxillaries.

2. Bypass vectored vents for auxillaries in pa_event(). This must be done by the
master vector hook, we can't afford a table that big.

3. Throw error on attempt to establish an auxillary vector in pa_eventover().

	modified:   portable/gnome_widgets.c

1. Define light and dark (off) focus widget reserved events.

2. In editbox_event(), send focus to parent as light or dark focus events.
See notes.

3. Implement focus highlighting in dropbox_draw().

4. In dropbox_event(), implement light and dark focus messages as focus on or
off followed by redraw.

5. In dropeditbox_event(), pass light and dark focus messages from edit box
widget to dropbox widget.

	modified:   tests/widget_test.c

Notes:

Getting the proper behavior from drop edit boxes was tricky. Attempts to pass
focus messages along didn't work, the rule there could be "one message source,
one or more receivers". The focus messages were fighting with the base
subsystem.

The system I used was to leave the focus to the edit widget, which is the one
that is actually getting focus (input characters), and treat the dropbox as just
getting "highlight the frame" messages, and thus no conflict.

The drawback to this is it needed user/module defined events, but it was
probably time to introduce that in any case. For future proofing, I left lots
of room for messages (4k each), and defined a block for each potential module.