Fix bug in scrollbars.
 Changes to be committed:
	modified:   portable/gnome_widgets.c

1. In both scrollvert_draw() and scrollhoriz_draw(), moved the bar click logic
back into scrollvert_event() and scrollhoriz_event(), respectively. The issue
we were detecting the click from the edge of pressed, which caused a loop, since
the program receiving the event moves the scrollbar to match, causing a redraw,
and then detects the click again, etc.

The answer is to use the fact that events are single shot. The drag does not
have this issue, but may refactor it later. The lesson is click handling should
be in events. This is the only widget that used that plan.

	modified:   tests/widget_test.c