Clean up call to addvect(); introduce full blank check to rotate.c.
 Changes to be committed:
	modified:   linux/graphics.c

1. Remove unnecessary parens.

	modified:   linux/rotated.c

1. Added blank(), check for all banks or null.

2. Refactor XRotDrawString() and xRotDrawAlignedString() to check blank or null
strings passed. This makes them a no-op, and the check avoids a crash later in
the routine when it passes 0 to X Windows routines.

There are certainly other cases where this is an issue, but these are the only
calls we make to the library. Note we have to leave background draw routines
alone since they may not be no-op cases (they draw backgrounds).